requirements.txt
azure-functions
pyodbc
python-dotenv
sqlalchemy
pytest
sqlalchemy
pydantic